--- a/.travis.yml
+++ b/.travis.yml
@@ -19,6 +19,3 @@ addons:
     packages:
     - g++-4.8
     - gcc-4.8
-
-before_install:
-  - npm install -g browserify standard