--- a/lib/createElement.js
+++ b/lib/createElement.js
@@ -13,10 +13,9 @@ var Element = require('./Element')
 module.exports = function createElement (name, attrs /*, child1, child2, ...*/) {
   var el = new Element(name, attrs)
 
-  var children = Array.prototype.slice.call(arguments, 2)
+  for (var i = 2; i < arguments.length; i++) {
+    el.cnode(arguments[i])
+  }
 
-  children.forEach(function (child) {
-    el.cnode(child)
-  })
   return el
 }