--- a/lib/tag.js
+++ b/lib/tag.js
@@ -5,13 +5,12 @@ var parse = require('./parse')
 
 module.exports = function tag (/* [literals], ...substitutions */) {
   var literals = arguments[0]
-  var substitutions = Array.prototype.slice.call(arguments, 1)
 
   var str = ''
 
-  for (var i = 0; i < substitutions.length; i++) {
-    str += literals[i]
-    str += escape(substitutions[i])
+  for (var i = 1; i < arguments.length; i++) {
+    str += literals[i - 1]
+    str += escape(arguments[i])
   }
   str += literals[literals.length - 1]