this is useful for es6 and you want to check something is an instance of Element but also for a more functional approach 

```javascript
import ltx from 'ltx'

ltx.is(foo)
```

instead of 

```javascript
import ltx from 'ltx'
const Element = {ltx}

foo instanceof Element || foo === 'string'
```