@astro 2016-06-07
Although enlarging the API surface, convenience is good.

Once at it, I was thinking about other static methods: how about making the default export a shortcut to the Element constructor, like:
```js
module.exports = createElement
```

Oh wait, that's already used for tagged template literals. Does my proposition conflict with that?