@sonnyp 2016-06-07
`module.exports` is tag indeed but since it's not documented si we could change that.

Although I believe the current behavior to be the most convenient for most usage.

```javasript
const xml = ltx('<some xml="without user input"/>')

const xml = ltx`
  <some xml="${with user input}"/>
`
```

we already have a `ltx.createElement` method which is JSX compatible

```javascript
// use ltx.createElement pragma
const xml = <some xml="{with user input}"/>
```