@sonnyp 2016-06-07
I've renamed `ltx.is` --> `ltx.isNode`