@sonnyp 2016-06-07
# benchmark

## master

```
suite ltx
tag with template literal x 62,935 ops/sec ±1.43% (86 runs sampled)
tag with direct call x 65,062 ops/sec ±1.48% (89 runs sampled)
serialize and parse x 76,365 ops/sec ±1.54% (85 runs sampled)
createElement (jsx) x 554,031 ops/sec ±2.01% (87 runs sampled)
```

# this
```
suite ltx
tag with template literal x 69,788 ops/sec ±1.58% (86 runs sampled)
tag with direct call x 71,470 ops/sec ±1.79% (87 runs sampled)
serialize and parse x 72,761 ops/sec ±1.48% (87 runs sampled)
createElement (jsx) x 2,366,923 ops/sec ±5.82% (86 runs sampled)
```